+++ src/libs/client/DeprecatedLoader.hpp
@@ -63,7 +63,7 @@ public:
 	string load_patch(const Glib::ustring&  filename,
 	                  boost::optional<Path> parent_path,
 	                  string                name,
-	                  size_t                poly,
+	                  uint32_t              poly,
 	                  MetadataMap           initial_data,
 	                  bool                  existing = false);
