+++ src/libs/engine/LADSPANode.hpp
@@ -34,7 +34,7 @@ namespace Ingen {
 class LADSPANode : public NodeBase
 {
 public:
-	LADSPANode(const Plugin* plugin, const string& name, size_t poly, Patch* parent, const LADSPA_Descriptor* descriptor, SampleRate srate, size_t buffer_size);
+	LADSPANode(const Plugin* plugin, const string& name, uint32_t poly, Patch* parent, const LADSPA_Descriptor* descriptor, SampleRate srate, size_t buffer_size);
 	virtual ~LADSPANode();
 
 	virtual bool instantiate();
@@ -44,7 +44,7 @@ public:
 	
 	void process(SampleCount nframes, FrameTime start, FrameTime end);
 	
-	void set_port_buffer(size_t voice, size_t port_num, Buffer* buf);
+	void set_port_buffer(uint32_t voice, uint32_t port_num, Buffer* buf);
 
 	const Plugin* plugin() const  { return _plugin; }
 	void plugin(const Plugin* const pi) { _plugin = pi; }
