+++ src/libs/engine/MidiControlNode.cpp
@@ -29,7 +29,7 @@
 namespace Ingen {
 
 	
-MidiControlNode::MidiControlNode(const string& path, size_t poly, Patch* parent, SampleRate srate, size_t buffer_size)
+MidiControlNode::MidiControlNode(const string& path, uint32_t poly, Patch* parent, SampleRate srate, size_t buffer_size)
 : InternalNode(new Plugin(Plugin::Internal, "ingen:control_node"), path, 1, parent, srate, buffer_size),
   _learning(false)
 {
