+++ src/libs/engine/OutputPort.hpp
@@ -42,8 +42,8 @@ class OutputPort : virtual public Port
 public:
 	OutputPort(Node*         parent,
 	           const string& name,
-	           size_t        index,
-	           size_t        poly,
+	           uint32_t      index,
+	           uint32_t      poly,
 	           DataType      type,
 	           size_t        buffer_size)
 	: Port(parent, name, index, poly, type, buffer_size)
